chore(ui): MetaNav adjustements
Die MetaNav wird jetzt innerhalb von MainNav.tsx definiert und auch dort in MainNav integriert, da MetaNav nur von MainNav genutzt wird, und da die MetaNav für Mobile innerhalb von MobileNav definiert werden soll. Der apps-Array wird vorübergehend in menuData definiert um für das MobileNav nutzbar zu sein. Perspektivisch soll apps im Config-Package definiert werden.
