+++ packages/ui/src/components/layouts/Container.tsx
@@ -1,6 +1,6 @@
+import { SiteHeader } from "@northware/ui/components/menu/SiteHeader";
 import { cn } from "@northware/ui/utils";
 import { ReactNode } from "react";
-import { SiteHeader } from "@northware/ui/components/menu/SiteHeader";
 
 export function Container({
   children,
@@ -10,9 +10,9 @@ export function Container({
   className?: string;
 }) {
   return (
-    <main className={cn("container", className)}>
+    <>
       <SiteHeader />
-      {children}
-    </main>
+      <main className={cn("container", className)}>{children}</main>
+    </>
   );
 }
